[refurb] Avoid FURB142 fix when loop variable is used after the loop
