+++ crates/ruff_linter/resources/test/fixtures/refurb/FURB142_shadow.py
@@ -1,0 +1,4 @@
+# Don't fix when loop variable shadows a non-LoopVar outer binding
+x = 0
+for x in (1, 2, 3):
+    s.add(x)
