+++ crates/ruff_linter/src/rules/refurb/rules/for_loop_writes.rs
@@ -1,11 +1,11 @@
 use ruff_macros::{ViolationMetadata, derive_message_formats};
-use ruff_python_ast::{Expr, ExprList, ExprName, ExprTuple, Stmt, StmtFor};
+use ruff_python_ast::{Expr, ExprName, Stmt, StmtFor};
 use ruff_python_semantic::analyze::typing;
-use ruff_python_semantic::{Binding, ScopeId, SemanticModel, TypingOnlyBindingsStatus};
-use ruff_text_size::{Ranged, TextRange, TextSize};
+use ruff_python_semantic::{Binding, ScopeId};
+use ruff_text_size::Ranged;
 
 use crate::checkers::ast::Checker;
-use crate::rules::refurb::helpers::IterLocation;
+use crate::rules::refurb::helpers::{IterLocation, binding_names, loop_variables_are_used_outside_loop};
 use crate::{AlwaysFixableViolation, Applicability, Edit, Fix};
 
 use crate::rules::refurb::helpers::parenthesize_loop_iter_if_necessary;
@@ -106,34 +106,6 @@ pub(crate) fn for_loop_writes_stmt(checker: &Checker, for_stmt: &StmtFor) {
     for_loop_writes(checker, for_stmt, scope_id, &[]);
 }
 
-/// Find the names in a `for` loop target
-/// that are assigned to during iteration.
-///
-/// ```python
-/// for ((), [(a,), [[b]]], c.d, e[f], *[*g]) in h:
-///     #      ^      ^                   ^
-///     ...
-/// ```
-fn binding_names(for_target: &Expr) -> Vec<&ExprName> {
-    fn collect_names<'a>(expr: &'a Expr, names: &mut Vec<&'a ExprName>) {
-        match expr {
-            Expr::Name(name) => names.push(name),
-
-            Expr::Starred(starred) => collect_names(&starred.value, names),
-
-            Expr::List(ExprList { elts, .. }) | Expr::Tuple(ExprTuple { elts, .. }) => elts
-                .iter()
-                .for_each(|element| collect_names(element, names)),
-
-            _ => {}
-        }
-    }
-
-    let mut names = vec![];
-    collect_names(for_target, &mut names);
-    names
-}
-
 /// FURB122
 fn for_loop_writes(
     checker: &Checker,
@@ -223,50 +195,4 @@ fn for_loop_writes(
             for_stmt.range,
         )
         .set_fix(fix);
-}
-
-fn loop_variables_are_used_outside_loop(
-    binding_names: &[&ExprName],
-    loop_range: TextRange,
-    semantic: &SemanticModel,
-    scope_id: ScopeId,
-) -> bool {
-    let find_binding_id = |name: &ExprName, offset: TextSize| {
-        semantic.simulate_runtime_load_at_location_in_scope(
-            name.id.as_str(),
-            TextRange::at(offset, 0.into()),
-            scope_id,
-            TypingOnlyBindingsStatus::Disallowed,
-        )
-    };
-
-    // If the load simulation succeeds at the position right before the loop,
-    // that binding is shadowed.
-    // ```python
-    //   a = 1
-    //   for a in b: ...
-    // # ^ Load here
-    // ```
-    let name_overwrites_outer =
-        |name: &ExprName| find_binding_id(name, loop_range.start()).is_some();
-
-    let name_is_used_later = |name: &ExprName| {
-        let Some(binding_id) = find_binding_id(name, loop_range.end()) else {
-            return false;
-        };
-
-        for reference_id in semantic.binding(binding_id).references() {
-            let reference = semantic.reference(reference_id);
-
-            if !loop_range.contains_range(reference.range()) {
-                return true;
-            }
-        }
-
-        false
-    };
-
-    binding_names
-        .iter()
-        .any(|name| name_overwrites_outer(name) || name_is_used_later(name))
 }
